+++ bankfeeds/src/sdk/companies.ts
@@ -315,7 +315,7 @@ export class Companies {
      * List companies
      *
      * @remarks
-     * Returns a list of your companies. The company schema contains a list of [connections](https://docs.codat.io/codat-api#/schemas/Connection) related to the company.
+     * Returns a list of your companies. The company schema contains a list of [connections](https://docs.codat.io/bank-feeds-api#/schemas/Connection) related to the company.
      */
     async list(
         req: operations.ListCompaniesRequest,