+++ bankfeeds/src/sdk/connections.ts
@@ -26,7 +26,7 @@ export class Connections {
      * @remarks
      * Creates a connection for the company by providing a valid `platformKey`.
      *
-     * Use the [List Integrations](https://docs.codat.io/codat-api#/operations/list-integrations) endpoint to access valid platform keys.
+     * Use the [List Integrations](https://docs.codat.io/bank-feeds-api#/operations/list-integrations) endpoint to access valid platform keys.
      */
     async create(
         req: operations.CreateDataConnectionRequest,