+++ bankfeeds/src/sdk/sdk.ts
@@ -53,8 +53,8 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.34.1";
-    genVersion = "2.86.6";
+    sdkVersion = "0.34.2";
+    genVersion = "2.86.10";
 
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);