+++ bankfeeds/RELEASES.md
@@ -627,3 +627,13 @@ Based on:
 - [typescript v0.34.1] bankfeeds
 ### Releases
 - [NPM v0.34.1] https://www.npmjs.com/package/@codat/bank-feeds/v/0.34.1 - bankfeeds
+
+## 2023-08-24 00:03:51
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.74.9 (2.86.10) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.34.2] bankfeeds
+### Releases
+- [NPM v0.34.2] https://www.npmjs.com/package/@codat/bank-feeds/v/0.34.2 - bankfeeds